+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 ### Bug Fixes
 - Fix unwanted localization in badge SVG generation (#629)
 - Update email template to handle not yet uploaded ping bodies
+- Add small delay in transports.Email.notify to allow ping body to upload
 
 ## v2.0.1 - 2022-03-18